requirements.txt
torch>=2.5.0
torchvision>=0.20.0
numpy>=1.26.0
scipy>=1.15.0
PyYAML>=6.0.0
tqdm>=4.67.0
matplotlib>=3.10.0

